refactor: move all file manipulation in to 1 place
A new Generator class to encapsulate all file manipulation.

BREAKING CHANGE: generate() is deprecated in favor of Generator
constructor.
