+++ test/generateFixture.ts
@@ -1,23 +1,15 @@
 import * as pathHelper from 'path'
-import generate from '../src/generate'
-import { accumulate } from '@johngw/async-iterator'
+import Generator from '../src/Generator'
 
 export default async function generateFixture(name: string, types: string[]) {
-  const [file] = await accumulate(
-    generate({
-      buildInstructions: [
-        {
-          targetFile: pathHelper.join(__dirname, `${name}.runtypes.ts`),
-          sourceTypes: [
-            {
-              file: pathHelper.join(__dirname, `${name}.ts`),
-              type: types,
-            },
-          ],
-        },
-      ],
-    })
-  )
+  const generator = new Generator({
+    targetFile: pathHelper.join(__dirname, `${name}.runtypes.ts`),
+  })
 
-  return file
+  return generator.generate([
+    {
+      file: pathHelper.join(__dirname, `${name}.ts`),
+      type: types,
+    },
+  ])
 }
