+++ test/minItems.test.ts
@@ -1,26 +1,17 @@
 import * as pathHelper from 'path'
-import { SourceFile } from 'ts-morph'
-import generate from '../src/generate'
+import Generator from '../src/Generator'
 
 test('json schema', async () => {
-  let file: SourceFile
+  const generator = new Generator({
+    targetFile: pathHelper.join(__dirname, `minItems.schema.runtypes.ts`),
+  })
 
-  for await (const _file of generate({
-    buildInstructions: [
-      {
-        targetFile: pathHelper.join(__dirname, `minItems.schema.runtypes.ts`),
-        sourceTypes: [
-          {
-            file: pathHelper.join(__dirname, 'minItems.schema.json'),
-            type: 'ExampleSchema',
-          },
-        ],
-      },
-    ],
-  })) {
-    file = _file
-    break
-  }
+  const file = await generator.generate([
+    {
+      file: pathHelper.join(__dirname, 'minItems.schema.json'),
+      type: 'ExampleSchema',
+    },
+  ])
 
   expect(file!.getText()).toMatchInlineSnapshot(`
 "import { Record, Tuple, Dictionary, String, Unknown, Static } from 'runtypes';
